pyproject.toml
[project]
name = "pydanticInput"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "M Laraib Ali", email = "laraib.ali@soco-engineers.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.7",
    "pyside6-essentials>=6.9.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 80
indent-width = 4
preview = true

[tool.ruff.lint]
isort.lines-after-imports = 2
isort.split-on-trailing-comma = false
select = [
    # "C",   # Complexity checks (e.g., McCabe complexity, comprehensions)
    "ANN001", "ANN201", "ANN401", # flake8-annotations (required strict type annotations for public functions)
    "S",   # flake8-bandit (checks basic security issues in code)
    "BLE", # flake8-blind-except (checks the except blocks that do not specify exception)
    "FBT", # flake8-boolean-trap (ensure that boolean args can be used with kw only)
    "E",   # pycodestyle errors (PEP 8 style guide violations)
    "W",   # pycodestyle warnings (e.g., extra spaces, indentation issues)
    "DOC", # pydoclint issues (e.g., extra or missing return, yield, warnings)
    "A",   # flake8-buitins (check variable and function names to not shadow builtins)
    "N",   # Naming convention checks (e.g., PEP 8 variable and function names)
    "F",   # Pyflakes errors (e.g., unused imports, undefined variables)
    "I",   # isort (Ensures imports are sorted properly)
    "B",   # flake8-bugbear (Detects likely bugs and bad practices)
    "TID", # flake8-tidy-imports (Checks for banned or misplaced imports)
    "UP",  # pyupgrade (Automatically updates old Python syntax)
    "YTT", # flake8-2020 (Detects outdated Python 2/3 compatibility issues)
    "FLY", # flynt (Converts old-style string formatting to f-strings)
    "PIE", # flake8-pie
    # "PL",  # pylint
    "RUF", # Ruff-specific rules (Additional optimizations and best practices)
]

ignore = ["S307"]

[tool.ruff.format]
docstring-code-format = true
skip-magic-trailing-comma = true